+++ Makefile
@@ -3,7 +3,7 @@ ldflags := -X "main.GitRivision=$(gitrev)"
 packages := $(shell go list ./... | grep -v memsniff/vendor)
 gometalinter := ${GOPATH}/bin/gometalinter.v1
 
-all: install lint test
+all: lint test
 
 install:
 	go install -x -ldflags "$(ldflags)" $(packages)
@@ -15,5 +15,5 @@ $(gometalinter):
 	go get -u gopkg.in/alecthomas/gometalinter.v1
 	gometalinter.v1 --install
 
-lint: $(gometalinter)
+lint: $(gometalinter) install
 	gometalinter.v1 --vendor --deadline 1m --enable-gc --disable=aligncheck ./...
